Make module-verison switch required for Azure ARM only
As the moduleVersion constant is no longer generated for data-plane, the
switch should only be required for Azure ARM.
