+++ src/autorest-configuration.md
@@ -72,7 +72,7 @@ help-content:
         type: boolean
         description: Indicates if generated clients are to be exported.  Default to true for ARM, false for data-plane.
       - key: module-version
-        description: Semantic version to include in generated telemetryInfo constant without the leading 'v' (e.g. 1.2.3).
+        description: When --azure-arm is true, semantic version to include in generated telemetryInfo constant without the leading 'v' (e.g. 1.2.3).
         type: string
       - key: group-parameters
         description: Enables parameter grouping via x-ms-parameter-grouping, defaults to true.
