+++ src/generator/constants.ts
@@ -10,8 +10,18 @@ import { length, values } from '@azure-tools/linq';
 import { contentPreamble, hasDescription, sortAscending } from './helpers';
 import { commentLength } from '../common/helpers';
 
+async function getModuleVersion(session: Session<CodeModel>): Promise<string> {
+  const version = await session.getValue('module-version', '');
+  if (version === '') {
+    throw new Error('--module-version is a required parameter');
+  } else if (!version.match(/^\d+\.\d+\.\d+$/) && !version.match(/^\d+\.\d+\.\d+-beta\.\d+$/)) {
+    throw new Error(`module version ${version} must in the format major.minor.patch[-beta.N]`);
+  }
+  return version;
+}
+
 // Creates the content in constants.go
-export async function generateConstants(session: Session<CodeModel>, version: string): Promise<string> {
+export async function generateConstants(session: Session<CodeModel>): Promise<string> {
   // lack of operation groups indicates model-only mode.
   if (length(session.model.operationGroups) === 0) {
     return '';
@@ -22,6 +32,7 @@ export async function generateConstants(session: Session<CodeModel>, version: st
   }
   // data-plane clients must manage their own constants for these values
   if (<boolean>session.model.language.go!.azureARM) {
+    const version = await getModuleVersion(session);
     text += `const (\n`;
     text += `\tmoduleName = "${session.model.language.go!.packageName}"\n`;
     text += `\tmoduleVersion = "v${version}"\n`;
